使用stdint.h

+++ src/1/1.1/main.c
@@ -3,7 +3,7 @@
 
 int main() {
   printf("你好，世界\n");
-  __int128_t num = 256;
+  int64_t num = 256;
   printBinary(&num, sizeof(num));
   return 0;
 }
